requirements.txt
# Requirements per GitHub Actions

# Web scraping
playwright==1.40.0
trafilatura==1.6.2
beautifulsoup4==4.12.2
lxml_html_clean==0.1.1
requests==2.31.0

# Database
supabase==2.3.0

# Utils
tqdm==4.66.0
python-dotenv==1.0.0
